requirements.txt
pandas
openpyxl
openai
requests
asyncio
djangorestframework